DEMOS-2385: broke this into manageable files
This overall script was huge and hard to read so i did some logical
breakdowns to make this fit.
